requirement.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
anndata==0.9.2
asciitree==0.3.3
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.3.0
backcall==0.2.0
Bottleneck==1.3.7
Brotli==1.0.9
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
cytoolz==0.12.3
dask==2023.5.0
dask-image==2023.3.0
debugpy==1.8.5
decorator==5.1.1
docrep==0.3.2
einops==0.8.0
elasticsearch==1.9.0
executing==2.1.0
fasteners==0.19
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
get-annotations==0.1.2
googledrivedownloader==0.4
gprofiler==1.2.2
gprofiler-official==1.0.0
grpcio==1.62.2
h5py==3.11.0
idna==3.7
igraph==0.11.5
imagecodecs-lite==2019.12.3
imageio==2.36.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflect==7.4.0
ipykernel==6.29.5
ipython==8.12.2
isodate==0.7.2
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
joypy==0.2.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
leidenalg==0.10.2
llvmlite==0.38.1
locket==1.0.0
louvain==0.8.2
Markdown==3.6
MarkupSafe==2.1.3
matplotlib==3.7.5
matplotlib-inline==0.1.7
matplotlib-scalebar==0.8.1
more-itertools==10.5.0
multidict==6.1.0
natsort==8.4.0
nest_asyncio==1.6.0
networkx==3.1
numba==0.55.2
numcodecs==0.12.1
numexpr==2.7.3
numpy==1.22.4
omnipath==1.0.8
opencv-python==4.10.0.84
packaging==24.1
pandas==2.0.3
parso==0.8.4
partd==1.4.1
patsy==1.0.1
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.3.0
PIMS==0.7
pip==24.2
platformdirs==3.10.0
pooch==1.7.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyDeprecate==0.3.2
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0
python-louvain==0.16
pytorch-lightning==1.6.0
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.2.0
rdflib==7.1.1
requests==2.32.3
s-dbw==0.4.0
scanpy==1.9.8
scikit-image==0.20.0
scikit-learn==1.3.0
scikit-misc==0.1.4
scipy==1.9.1
scprep==1.2.3
seaborn==0.13.2
session_info==1.0.0
setuptools==59.5.0
six==1.16.0
slicerator==1.1.0
slingshot==0.0.5
squidpy==1.2.3
stack-data==0.6.2
statsmodels==0.14.1
stdlib-list==0.10.0
tensorboard==2.17.1
tensorboard-data-server==0.7.0
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2020.6.3
toolz==1.0.0
torch==1.8.0
torch-cluster==1.5.9
torch-geometric==2.0.0
torch-scatter==2.0.8
torch-sparse==0.6.12
torch-spline-conv==1.2.1
torchaudio==0.8.0a0+a751e1d
torchmetrics==0.4.1
torchvision==0.9.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typeguard==4.4.0
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
urllib3==1.26.20
validators==0.34.0
wcwidth==0.2.13
Werkzeug==3.0.6
wheel==0.44.0
wrapt==1.17.2
xarray==2023.1.0
yacs==0.1.6
yarl==1.15.2
zarr==2.16.1
zipp==3.20.2
